Create uEditHDForcePreviewModeOnInit
By default, uEdit HD shows the XLIM documents in "Normal" mode: variable text areas come with a blue highlight and a dotted border; variable images come with a dotted border and a red text atop, with the image resize mode; while this is useful to regular InDesign users (as the same features also exist in uCreate Designer), this may be surprising or disturbing to new uEdit HD users. Switching uEdit HD display mode to "Preview" instead of "Normal" (which is the purpose of this JavaScript) may be an option.